Create a folder for a new project
This project is called 'Simple blog page' where I
practice making a blog page for a website.
I created a directory for the webpage that includes
the index page and css file.
The index file contains simple boilerplate code and
the general article for the website.
The css file only contains an import link for the
font to be used.